refactor(mail): merge interdependent priority concerns
